requirements.txt
astroid==2.4.2
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.10
django-filter==2.2.0
djangorestframework==3.10.2
gunicorn==20.0.4
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
pycparser==2.20
pydub==0.24.1
pylint==2.5.3
pytz==2019.2
requests==2.22.0
six==1.15.0
SoundFile==0.10.3.post1
sqlparse==0.3.0
toml==0.10.1
uritemplate==3.0.0
urllib3==1.25.3
whitenoise==5.1.0
wrapt==1.12.1
